v0.3

+++ 21451121谢伟军/ProjectFinal/CDSideBarController/CDSideBarController.m
@@ -96,7 +96,7 @@
 - (void)onMenuButtonClick:(UIButton*)button
 {
     if ([self.delegate respondsToSelector:@selector(menuButtonClicked:)])
-        [self.delegate menuButtonClicked:button.tag];
+        [self.delegate menuButtonClicked:(int)button.tag];
     [self dismissMenuWithSelection:button];
 }
